Fixing integer division.

+++ IP_ledge_detect.cpp
@@ -64,11 +64,11 @@ int detect_lines(Mat src)
     {
         Vec4i l = lines[i];
         if (!(l[0] - l[2] == 0) && // If the line is NOT purely vertical (div by zero protection) and...
-           (((l[1] - l[3]) / (l[0] - l[2])) < ABS_MAX_LEDGE_SLOPE) && // If the slope of the line is less than 1/2 and...
-           (((l[1] - l[3]) / (l[0] - l[2])) > -ABS_MAX_LEDGE_SLOPE)) // If the slope of the line is greater than 1/2.
+           (((float)(l[1] - l[3]) / (float)(l[0] - l[2])) < ABS_MAX_LEDGE_SLOPE) && // If the slope of the line is less than 1/2 and...
+           (((float)(l[1] - l[3]) / (float)(l[0] - l[2])) > -ABS_MAX_LEDGE_SLOPE)) // If the slope of the line is greater than 1/2.
         {
             #if (1)
-            cout << "(" << l[0] << ", " << l[1] << "), (" << l[2] << ", " << l[3] << ") Slope = " << ((l[1] - l[3]) / (l[0] - l[2])) << endl;
+            cout << "(" << l[0] << ", " << l[1] << "), (" << l[2] << ", " << l[3] << ") Slope = " << ((float)(l[1] - l[3]) / (float)(l[0] - l[2])) << endl;
             #endif
             // only draw lines with slope less than 1/2
             line( cdst, Point(l[0], l[1]), Point(l[2], l[3]), Scalar(0,255,0), 1, CV_AA);
